requirements.txt
mutagen
pygame
Pillow